pyproject.toml
[project]
name = "AlphaPROBE"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "annoymous", email = "annoymous@example_mail.com"},
]
dependencies = [
    "numpy>=2.3.2",
    "torch @ https://download.pytorch.org/whl/cu121/torch-2.4.0%2Bcu121-cp311-cp311-linux_x86_64.whl",
    "torchvision @ https://download.pytorch.org/whl/cu121/torchvision-0.19.0%2Bcu121-cp311-cp311-linux_x86_64.whl",
    "torchaudio @ https://download.pytorch.org/whl/cu121/torchaudio-2.4.0%2Bcu121-cp311-cp311-linux_x86_64.whl",
    "jinja2>=3.1.6",
    "gymnasium>=1.2.0",
    "stable-baselines3>=2.7.0",
    "pandera>=0.25.0",
    "pyqlib>=0.9.6",
    "baostock>=0.8.9",
    "sb3-contrib>=2.7.0",
    "opencv-python>=4.11.0.86",
    "joblib==1.3.2",
    "tensorboard>=2.20.0",
    "seaborn>=0.13.2",
    "torchgfn>=1.2.1",
    "torch-geometric>=2.6.1",
    "xgboost>=3.0.5",
    "yahooquery>=2.4.1",
    "backoff==2.2.1",
    "retry==0.9.2",
    "transformers==4.56.0",
    "openai==1.55.1",
    "retrying==1.3.4",
    "httpx==0.27.2",
    "sentence-transformers==3.3.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true
